Removed autopep8 since it did not format python file added is file check in file & json operations added some minor pylinting errors

+++ py_file_operations/file_operations.py
@@ -4,82 +4,86 @@ import codecs
 import os
 import struct
 import shutil
+path = os.path
 
 
 class FileOperations:
     """File operations class """
     file_handle: object | None
-    file_data: str | bytes # type: ignore
+    file_data: str | bytes  # type: ignore
 
-    def open_file(self, file_name : str, opening_mode: str) -> None:
+    def open_file(self, file_name: str, opening_mode: str) -> None:
         """ Open file method """
         # Opening a file explicitely for reading purpose
         try:
-            with open(file_name, opening_mode, encoding='UTF-8') as handle:
-                self.file_handle = handle
+            if path.isfile(file_name):
+                with open(file_name, opening_mode, encoding='UTF-8') as handle:
+                    self.file_handle = handle
         except FileNotFoundError as file_not_found_error:
             print(file_not_found_error)
 
-
     def close_file(self) -> None:
         """Closing file explicitely"""
-        if not self.file_handle is None and self.file_handle.closed is False: # type: ignore
-            self.file_handle.close() # type: ignore
-
+        if not self.file_handle is None and self.file_handle.closed is False:  # type: ignore
+            self.file_handle.close()  # type: ignore
 
     def read_file(self, file_name: str) -> None:
         """ Reading file data method """
         try:
-            # with will take care of file to be closed after use.
-            with open(file_name,'r', encoding='UTF-8') as self.file_handle:
-                self.file_data = self.file_handle.read() # type: ignore
+            if path.isfile(file_name):
+                # with will take care of file to be closed after use.
+                with open(file_name, 'r', encoding='UTF-8') as self.file_handle:
+                    self.file_data = self.file_handle.read()  # type: ignore
         except FileNotFoundError as file_not_found_error:
             print(file_not_found_error)
-
 
     def read_binarydata_from_file(self, file_name: str) -> None:
         """ Reading binary data from file method """
+        binary_data_one: str = ''
+        binary_data_two: str  = ''
+        binary_data_three: str = ''
         try:
-            with open(file_name,'rb') as handle:
-                self.file_data: bytes = handle.read(8)
-                binary_data_one , binary_data_two, binary_data_three = struct.unpack('>hhl',
-                                                                                     self.file_data)
-                print('Binary data from file: ', binary_data_one, binary_data_two,
-                      binary_data_three)
+            if path.isfile(file_name):
+                with open(file_name, 'rb') as handle:
+                    self.file_data: bytes = handle.read(8)
+                    binary_data_one, binary_data_two, binary_data_three = struct.unpack('>hhl',
+                    self.file_data)
+                    print('Binary data from file: ', binary_data_one,
+                          binary_data_two, binary_data_three)
         except FileNotFoundError as file_not_found_error:
             print(file_not_found_error)
-
 
     def write_file(self, file_name: str) -> None:
         """ Write operation currently causing an error. DO NOT USE. """
         # # w+ Opens a disk file for updating(reading and writing)
         try:
-            with open(file_name,'w+', encoding='utf-8') as handle:
-                self.file_handle = handle
-                data: str = 'This is a file write operation.'
-                self.file_handle.write(data)
+            if path.isfile(file_name):
+                with open(file_name, 'w+', encoding='utf-8') as handle:
+                    self.file_handle = handle
+                    data: str = 'This is a file write operation.'
+                    self.file_handle.write(data)
         except FileNotFoundError as file_not_found_error:
             print(file_not_found_error)
-
 
     def append_file(self, file_name: str) -> None:
         """Appending file contents method"""
         # # a Opens a file to append the contents to the end of the file.
         try:
-            with open(file_name,'a+', encoding='utf-8') as self.file_handle:
-                data: str = 'Appending file contents. '
-                self.file_handle.write(data)
-                self.close_file()
+            if path.isfile(file_name):
+                with open(file_name, 'a+', encoding='utf-8') as self.file_handle:
+                    data: str = 'Appending file contents. '
+                    self.file_handle.write(data)
+                    self.close_file()
         except FileNotFoundError as file_not_found_error:
             print(file_not_found_error)
-
 
     def delete_file(self, file_name: str) -> None:
         """ Deleting file method """
         try:
-            if os.access(file_name, os.F_OK):
-                os.remove(file_name)
-                print(file_name, ' deleted successfully. ')
+            if path.isfile(file_name):
+                if os.access(file_name, os.F_OK):
+                    os.remove(file_name)
+                    print(file_name, ' deleted successfully. ')
         except FileNotFoundError as file_not_found_error:
             print(file_not_found_error)
         except FileExistsError as file_exists_error:
@@ -87,40 +91,37 @@ class FileOperations:
         except OSError as os_error:
             print(os_error)
 
-
-    def rename_file(self, oldfile_name : str, newfile_name: str) -> None:
+    def rename_file(self, oldfile_name: str, newfile_name: str) -> None:
         """ Renaming file method - causing Access is denied error DO NOT USE """
         try:
-            if os.access(oldfile_name, os.F_OK):
-                os.rename(oldfile_name, newfile_name)
-                print('File has been renamed successfully. ')
+            if path.isfile(oldfile_name):
+                if os.access(oldfile_name, os.F_OK):
+                    os.rename(oldfile_name, newfile_name)
+                    print('File has been renamed successfully. ')
         except FileNotFoundError as file_not_found_error:
             print(file_not_found_error)
         except OSError as os_error:
             print(os_error)
 
-
-    def file_copy(self, source_file_name : str, dest_file_name: str) -> None:
+    def file_copy(self, source_file_name: str, dest_file_name: str) -> None:
         """ File copy method """
         try:
-            result: str = shutil.copyfile(source_file_name, dest_file_name)
-            print(result)
+            if path.isfile(source_file_name):
+                result: str = shutil.copyfile(source_file_name, dest_file_name)
+                print(result)
         except shutil.SameFileError as same_file_error:
             print(same_file_error)
         except shutil.SpecialFileError as special_file_error:
             print(special_file_error)
-
 
     def file_wildcards(self) -> None:
         """ File wild cards method """
         files_list: list[str] = glob.glob('*.py')
         print(files_list)
 
-
     def display_file_data(self) -> str | bytes | None:
         """ Returning retrieved file data method """
         return self.file_data
-
 
     def convert_between_file_encoding(self, file_name: str) -> None:
         """ Convert between file encoding method """
@@ -128,21 +129,18 @@ class FileOperations:
         # reading data & called bytes for the conversion. " DO NOT USE
         # file_handle causing conversion error TextIOWrapper to codecs._Stream
         try:
-            with open(file_name, 'r', encoding='latin-1' ) as self.file_handle:
-                newfile_handle =  codecs.StreamRecoder(self.file_handle, # type: ignore
-                                                       codecs.getencoder('utf-8'),
-                                                       codecs.getdecoder('utf-8'),
-                                                       codecs.getreader('latin-1'),
-                                                       codecs.getwriter('latin-1'),
-                                                       errors='strict')
-                data = newfile_handle.read().decode(
-                    encoding='iso-8859-1').encode(
-                    encoding='utf-8')
-                print(data)
+            if path.isfile(file_name):
+                with open(file_name, 'r', encoding='latin-1') as self.file_handle:
+                    newfile_handle = codecs.StreamRecoder(self.file_handle,  # type: ignore
+                    codecs.getencoder('utf-8'), codecs.getdecoder('utf-8'),
+                    codecs.getreader('latin-1'), codecs.getwriter('latin-1'),
+                    'strict') # type: ignore
+                    data = newfile_handle.read().decode(encoding='iso-8859-1').encode(
+                        encoding='utf-8')
+                    print(data)
 
         except OSError as os_error:
             print(os_error)
-
 
     # def convert_latin_data(self):
     #     try:
@@ -156,6 +154,7 @@ class FileOperations:
     #     except Exception as ex:
     #         print(ex)
 
+
 if __name__ == '__main__':
     try:
         # Creating file operation class object.
@@ -167,7 +166,7 @@ if __name__ == '__main__':
         old_file_name: str = 'C:\\Data\\tempfile.txt'
         new_file_name: str = 'C:\\Data\\tempfile_one.txt'
 
-        file_instance.rename_file(old_file_name, new_file_name )
+        file_instance.rename_file(old_file_name, new_file_name)
 
         latin_data_file_name: str = 'C:\\Data\\latindata.txt'
         # file_instance.convert_between_file_encoding(latin_data_file_name)
@@ -207,7 +206,7 @@ if __name__ == '__main__':
         file_instance.read_file(txt_file_name)
 
         # Calling file display data
-        file_contents  = file_instance.display_file_data()
+        file_contents = file_instance.display_file_data()
 
         # Printing file contents after reading.
         print('Reading after appending file contents: ', file_contents)
@@ -215,8 +214,8 @@ if __name__ == '__main__':
         # Calling file close explicitely if file has not been closed.
         file_instance.close_file()
 
-        source_file : str = 'C:\\Data\\sourcefile.txt'
-        dest_file : str = 'C:\\Data\\destinationfile.txt'
+        source_file: str = 'C:\\Data\\sourcefile.txt'
+        dest_file: str = 'C:\\Data\\destinationfile.txt'
         file_instance.file_copy(source_file,  dest_file)
 
         del_file_name: str = 'C:\\Data\\delfile.txt'
